CaptureActivity code detach
